fix: address remaining Qodo review feedback (#6-#11)
- Move DeviceLegendGroup to its own file (one class per file rule)
- Switch SetCursor from Dispatcher.Invoke to BeginInvoke for non-blocking UI
- Cache last cursor to avoid redundant updates on mouse move
- Add null/empty/zero input guards to MinMaxDownsampler.Downsample
- Guard GetMinimapDataRange against NaN/Infinity/zero returns
