+++ Daqifi.Desktop/Loggers/DatabaseLogger.cs
@@ -23,34 +23,6 @@ using FontWeights = OxyPlot.FontWeights;
 
 namespace Daqifi.Desktop.Logger;
 
-/// <summary>
-/// Groups legend items by device for compact display in the legend panel.
-/// </summary>
-public partial class DeviceLegendGroup : ObservableObject
-{
-    /// <summary>
-    /// Full device serial number.
-    /// </summary>
-    public string DeviceSerialNo { get; }
-
-    /// <summary>
-    /// Truncated serial for display (e.g., "...4104").
-    /// </summary>
-    public string TruncatedSerialNo => DeviceSerialNo?.Length > 4
-        ? $"...{DeviceSerialNo[^4..]}"
-        : DeviceSerialNo ?? string.Empty;
-
-    /// <summary>
-    /// Channel legend items belonging to this device.
-    /// </summary>
-    public ObservableCollection<LoggedSeriesLegendItem> Channels { get; } = new();
-
-    public DeviceLegendGroup(string deviceSerialNo)
-    {
-        DeviceSerialNo = deviceSerialNo;
-    }
-}
-
 public partial class LoggedSeriesLegendItem : ObservableObject
 {
     [ObservableProperty]
